There is a need to let the user choose which stories to show in the homepage banner.
Using the ACF v4 plugin I have  added a Relation field showing all the stories in the system.
The problem with this solution was that ACF v4 loads all the stories when a user editing a page,
which causes the page to crush due to large number of posts.
The soultion was to install ACF v5 PRO, in which the posts loading done by ajax in lazy load.
The new ACF plugin broke some of the current code, hence there was a need to adjust the files to the new version.

1) functions.php: Added a call to acf.php which includes all releated function for the ACF plugin.
2) acf.php: added a function for displaying the options page (was done automaticlly in v4) and a backwards compability function.
3) backend.php AND backend_repeater: fixed the backend custom repeater  problems, such as the inabilty to delete a quote \ dictionary term.
4) layout-banner: Made the banner to read the user custom stories, if he choose not show random stories.